requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.29
python-dotenv==1.0.0
pydantic==1.10.13
httpx==0.24.1
pg8000==1.30.5
anyio==3.7.1
supabase==1.0.4